Devide_screen
스크린 분리 및 모달 테스트
